Added OSThread.

+++ MyClient/src/main/java/com/yuyijq/client/MyThread.java
@@ -6,5 +6,5 @@ package com.yuyijq.client;
  * Time: 11:02 PM
  */
 public interface MyThread {
-    void sleep(int interval);
+    void sleep(Long interval);
 }
